合并原作者的更新

+++ src/main/java/com/thinkgem/jeesite/modules/sys/web/DictController.java
@@ -66,6 +66,10 @@ public class DictController extends BaseController {
 	@RequiresPermissions("sys:dict:edit")
 	@RequestMapping(value = "save")//@Valid 
 	public String save(Dict dict, Model model, RedirectAttributes redirectAttributes) {
+		if(Global.isDemoMode()){
+			addMessage(redirectAttributes, "演示模式，不允许操作！");
+			return "redirect:"+Global.getAdminPath()+"/sys/dict/?repage&type="+dict.getType();
+		}
 		if (!beanValidator(model, dict)){
 			return form(dict, model);
 		}
@@ -77,6 +81,10 @@ public class DictController extends BaseController {
 	@RequiresPermissions("sys:dict:edit")
 	@RequestMapping(value = "delete")
 	public String delete(Long id, RedirectAttributes redirectAttributes) {
+		if(Global.isDemoMode()){
+			addMessage(redirectAttributes, "演示模式，不允许操作！");
+			return "redirect:"+Global.getAdminPath()+"/sys/dict/?repage";
+		}
 		dictService.delete(id);
 		addMessage(redirectAttributes, "删除字典成功");
 		return "redirect:"+Global.getAdminPath()+"/sys/dict/?repage";
